+++ packages/react/src/components/DataTable/src/components/DataTableDropdown/index.ts
@@ -1,0 +1,5 @@
+import { DataTableDropdown } from "./DataTableDropdown";
+
+export { DataTableDropdown } from "./DataTableDropdown";
+export type { DataTableDropdownProps } from "./dataTableDropdown.types";
+export default DataTableDropdown;