+++ packages/react/src/components/DataTable/src/components/DataTableDropdownAction/index.ts
@@ -1,0 +1,5 @@
+import { DataTableDropdownAction } from "./DataTableDropdownAction";
+
+export { DataTableDropdownAction } from "./DataTableDropdownAction";
+export type { DataTableDropdownActionProps } from "./dataTableDropdownAction.types";
+export default DataTableDropdownAction;