+++ packages/react/src/components/DataTable/src/components/DataTableDropdownDivider/index.ts
@@ -1,0 +1,4 @@
+import { DataTableDropdownDivider } from "./DataTableDropdownDivider";
+
+export { DataTableDropdownDivider } from "./DataTableDropdownDivider";
+export default DataTableDropdownDivider;