+++ packages/react/src/components/DataTable/src/components/DataTableDropdownSection/index.ts
@@ -1,0 +1,1 @@
+export { DataTableDropdownSection } from "./DataTableDropdownSection";